Oprettet user-with-timeentries.html og slettet Sofies døde row mapper + min egen døde DateHandler-klasse
