+++ src/main/java/com/example/estimationtool/toolbox/dto/TaskWithTimeEntriesDTO.java
@@ -1,7 +1,7 @@
 package com.example.estimationtool.toolbox.dto;
 
 import com.example.estimationtool.model.Task;
-import com.example.estimationtool.model.timeEntry.TimeEntry;
+import com.example.estimationtool.model.TimeEntry;
 import java.util.List;
 
 public record TaskWithTimeEntriesDTO(Task task, List<TimeEntry> timeEntries) {
