+++ src/main/resources/templates/subtask/subtask-detail.html
@@ -34,6 +34,13 @@
     </div>
 </section>
 
+<!-- Knap der fører til subtaskens time entry-liste -->
+<div>
+    <a th:href="@{/subtasks/{subTaskId}/timeentries(subTaskId=${subtask.subTaskId})}">
+        <button>Se tilknyttede time entries</button>
+    </a>
+</div>
+
 <!-- Vis tilknyttet bruger -->
 <section class="mini-info">
     <h2>Tildelt bruger</h2>
