+++ src/main/resources/templates/task/task-detail.html
@@ -44,6 +44,13 @@
         </a>
     </div>
 
+    <!-- Til taskens time entry-liste -->
+    <div>
+        <a th:href="@{/tasks/{taskId}/timeentries(taskId=${task.taskId})}">
+            <button>Se tilknyttede time entries</button>
+        </a>
+    </div>
+
     <!-- Se tilknyttede brugere til task -->
     <div>
         <a th:href="@{/tasks/{taskId}/users(taskId=${task.taskId})}">
