+++ src/main/resources/templates/user/user-with-timeentries.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>User With Time Entries</title>
+</head>
+<body>
+
+</body>
+</html>
